refactor: also use SFC Pager component in ControlsHeader
