+++ server/zanata-frontend/src/app/editor/components/Pager/index.js
@@ -1,6 +1,6 @@
 // @ts-nocheck
 /*
- * Copyright 2016, Red Hat, Inc. and individual contributors as indicated by the
+ * Copyright 2018, Red Hat, Inc. and individual contributors as indicated by the
  * @author tags. See the copyright.txt file in the distribution for a full
  * listing of individual contributors.
  *
@@ -23,14 +23,6 @@
 import { Icon } from '../../../components'
 import React from 'react'
 import * as PropTypes from 'prop-types'
-import { connect } from 'react-redux'
-import {
-  firstPage,
-  nextPage,
-  previousPage,
-  lastPage
-} from '../../actions/controls-header-actions'
-import { getMaxPageIndex } from '../../selectors'
 
 /**
 * Paging buttons and current-page indicator.
@@ -74,41 +66,32 @@ export const Pager = ({
   nextPage,
   lastPage,
   pageCount,
-  pageNumber,
-  gettextCatalog
+  pageNumber
 }) => {
-  /* FIXME make this string translatable (was using angular gettext)
-  const pageXofY = pageCount
-    ? gettextCatalog.getString(
-      '{{currentPage}} of {{totalPages}}', {
-        currentPage: pageNumber,
-        totalPages: pageCount
-      })
-    : pageNumber*/
   const pageXofY = `${pageNumber} of ${pageCount}`
 
   const buttons = {
     first: {
       icon: 'previous',
-      title: gettextCatalog.getString('First page'),
+      title: 'First page',
       action: firstPage,
       disabled: pageNumber <= 1
     },
     prev: {
       icon: 'chevron-left',
-      title: gettextCatalog.getString('Previous page'),
+      title: 'Previous page',
       action: previousPage,
       disabled: pageNumber <= 1
     },
     next: {
       icon: 'chevron-right',
-      title: gettextCatalog.getString('Next page'),
+      title: 'Next page',
       action: nextPage,
       disabled: pageNumber >= pageCount
     },
     last: {
       icon: 'next',
-      title: gettextCatalog.getString('Last page'),
+      title: 'Last page',
       action: lastPage,
       disabled: pageNumber >= pageCount
     }
@@ -135,32 +118,7 @@ Pager.propTypes = {
   nextPage: PropTypes.func.isRequired,
   lastPage: PropTypes.func.isRequired,
   pageNumber: PropTypes.number.isRequired,
-  pageCount: PropTypes.number,
-
-  // DO NOT RENAME, the translation string extractor looks specifically
-  // for gettextCatalog.getString when generating the translation template.
-  gettextCatalog: PropTypes.shape({
-    getString: PropTypes.func.isRequired
-  }).isRequired
+  pageCount: PropTypes.number
 }
 
-const mapStateToProps = (state) => {
-  const { phrases, ui: { gettextCatalog } } = state
-  const pageCount = getMaxPageIndex(state) + 1
-  const pageNumber = Math.min(pageCount, phrases.paging.pageIndex + 1)
-
-  return {
-    pageCount: pageCount,
-    pageNumber: pageNumber,
-    gettextCatalog
-  }
-}
-
-const mapDispatchToProps = {
-  firstPage,
-  previousPage,
-  nextPage,
-  lastPage
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(Pager)
+export default Pager
